[FIX/#109] Diagnosis / AI 서버로 부터 response 받기 위한 수정

+++ src/main/java/capstone/capstone7/domain/diagnosis_result/dto/DiagnosisResultFromAIServer.java
@@ -10,5 +10,5 @@ public class DiagnosisResultFromAIServer {
     private String disease_name;
     private String in_info;
     private String out_info;
-    private List<Double> class_prob_list;
+    private List<Float> class_prob_list;
 }
